+++ apps/www/src/app/(main)/(docs)/components/input/examples/outlined-all-props.tsx
@@ -4,18 +4,22 @@ import { EmailIcon } from "@jamsr-ui/shared-icons";
 export const InputOutlinedAllProps = () => {
   return (
     <div className="flex flex-col gap-4">
-      <Input label="Enter your email (small)" size="sm" variant="outlined" />
-      <Input label="Enter your email (middle)" size="md" variant="outlined" />
-      <Input label="Enter your email (large)" size="lg" variant="outlined" />
-      <Input label="Enter your email" isFilled variant="outlined" />
-      
+      <Input label="Outlined Input (small)" size="sm" variant="outlined" />
+      <Input label="Outlined Input (medium)" size="md" variant="outlined" />
+      <Input label="Outlined Input (large)" size="lg" variant="outlined" />
+      <Input label="Outlined Input (Filled)" isFilled variant="outlined" />
       <Input
-        placeholder="Enter your email"
-        label="Enter your email"
+        label="Outlined Input"
+        placeholder="Input Placeholder"
+        variant="outlined"
+      />
+      <Input placeholder="Without Label" variant="outlined" />
+      <Input
+        placeholder="With start and end content"
+        label="Outlined Input"
         startContent={<EmailIcon />}
         endContent={<EmailIcon />}
         variant="outlined"
-        size="lg"
       />
     </div>
   );