+++ apps/www/src/app/(main)/(docs)/components/input/examples/standard-all-props.tsx
@@ -1,19 +1,25 @@
 import { Input } from "@jamsr-ui/react";
 import { EmailIcon } from "@jamsr-ui/shared-icons";
 
-export const InputOutlinedAllProps = () => {
+export const InputStandardAllProps = () => {
   return (
     <div className="flex flex-col gap-4">
-      <Input label="Enter your email" size="sm" variant="outlined" />
-      <Input label="Enter your email" size="md" variant="outlined" />
-      <Input label="Enter your email" size="lg" variant="outlined" />
-      <Input label="Enter your email" isFilled variant="outlined" />
+      <Input label="Standard Input (Small)" size="sm" variant="standard" />
+      <Input label="Standard Input (Medium)" size="md" variant="standard" />
+      <Input label="Standard Input (Large)" size="lg" variant="standard" />
+      <Input label="Standard Input (Filled)" isFilled variant="standard" />
       <Input
-        placeholder="Enter your email"
-        label="Enter your email"
+        label="Standard Input (Placeholder)"
+        placeholder="Standard Input Placeholder"
+        variant="standard"
+      />
+      <Input placeholder="Without Label" variant="standard" />
+      <Input
+        placeholder="Standard Input"
+        label="End and start content"
         startContent={<EmailIcon />}
         endContent={<EmailIcon />}
-        variant="outlined"
+        variant="standard"
       />
     </div>
   );