+++ apps/www/src/app/(main)/(docs)/components/input/examples/usage.tsx
@@ -1,5 +1,12 @@
 import { Input } from "@jamsr-ui/react";
 
 export const InputUsage = () => {
-  return <Input label="Enter your email" type="email" />;
+  return (
+    <div>
+      <Input label="Enter your email" type="email" />
+      <fieldset className="border-2 border-default-200">
+        <legend className="px-2">Fieldset</legend>
+      </fieldset>
+    </div>
+  );
 };