+++ apps/www/src/app/(main)/(home)/_components/hero/github-stars-count.tsx
@@ -1,6 +1,0 @@
-import { getGithubStarsCount } from "../../actions";
-
-export const GithubStarsCount = async () => {
-  const count = await getGithubStarsCount();
-  return count;
-};