+++ apps/www/src/app/(main)/(home)/actions.ts
@@ -3,12 +3,25 @@ export const getGithubStarsCount = async () => {
     const response = await fetch(
       "https://api.github.com/repos/jamsrworld/jamsr-ui",
       {
-        cache: "no-store",
+        next: {
+          revalidate: 60,
+        },
       },
     );
     const data = (await response.json()) as { stargazers_count: number };
     return data.stargazers_count;
   } catch {
     return 0;
+  }
+};
+
+export const getLatestVersion = async () => {
+  try {
+    const url = "https://registry.npmjs.org/@jamsr-ui/react/latest";
+    const response = await fetch(url);
+    const data = (await response.json()) as { version: string };
+    return data.version;
+  } catch {
+    return "0.0.0";
   }
 };