+++ apps/www/src/types/variants.d.ts
@@ -2,7 +2,7 @@ export type VariantTypes = {
   "accordion": ["custom-icon", "disabled", "end-content-outside", "end-content", "multiple-selection", "radius", "start-content-outside", "start-content", "subheading", "usage", "variants", "without-icon"];
   "alert": ["custom-icon", "radius", "status", "usage", "variants", "with-action", "with-description", "without-icon"];
   "autocomplete": ["choose-country", "controlled", "disabled-items", "disabled", "error-state", "helper-text", "multiple-controlled", "multiple-custom-render", "multiple", "start-end-content", "usage"];
-  "avatar": ["avatar.d", "bordered", "colors", "fallback-color", "fallbacks", "image-fallback", "nextjs-usage", "radius", "sizes", "usage"];
+  "avatar": ["bordered", "colors", "fallback-color", "fallbacks", "image-fallback", "nextjs-usage", "radius", "sizes", "usage"];
   "breadcrumbs": [];
   "button": ["colors", "disabled", "loading", "radius", "sizes", "usage", "variants-colors", "variants", "with-icons"];
   "card": ["radius", "start-end-content", "usage", "variants", "with-form"];
@@ -25,7 +25,7 @@ export type VariantTypes = {
   "file-upload-single": ["avatar", "controlled", "customized", "default-value", "disabled", "radius", "usage"];
   "header": ["hide-on-scroll", "static", "usage", "with-border"];
   "icon-button": ["colors", "disabled", "loading", "radius", "sizes", "usage", "variants-colors", "variants"];
-  "input": ["clearable", "controlled", "disabled", "error-state", "filled", "helper-text", "label-helper", "numbers-only", "optional", "outlined-all-props", "placeholder", "radius", "required", "secured-text", "sizes", "standard-all-props", "usage", "variants", "with-end-content", "with-icon", "with-placeholder", "with-start-content", "without-label"];
+  "input": ["bordered-all-props", "clearable", "controlled", "disabled", "error-state", "filled", "helper-text", "label-helper", "numbers-only", "optional", "outlined-all-props", "placeholder", "radius", "required", "secured-text", "sizes", "standard-all-props", "underlined-all-props", "usage", "variants", "with-end-content", "with-icon", "with-placeholder", "with-start-content", "without-label"];
   "kbd": ["keys", "radius", "usage"];
   "linear-progress": ["colors", "controlled", "usage"];
   "link": ["nextjs-usage", "underline", "usage"];