+++ packages/components/autocomplete/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @jamsr-ui/autocomplete
 
+## 0.0.130
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/input@0.0.121
+
 ## 0.0.129
 
 ### Patch Changes