+++ packages/components/input/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/input
 
+## 0.0.121
+
+### Patch Changes
+
+- feat: input variant
+
 ## 0.0.120
 
 ### Patch Changes